requirements.txt
Flask>=3.0
pymongo>=4.6
PyYAML>=6.0
matplotlib>=3.8
requests>=2.31
httpx>=0.27
